requirements.txt
DJANGO==2.1.3
PILLOW==5.3.0
django-crispy-forms==1.7.2